% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{Get_cell_type_score}
\alias{Get_cell_type_score}
\title{Cell type score}
\usage{
Get_cell_type_score(object, cell_type_label = "cell_type_overall")
}
\arguments{
\item{object}{= seurat object}

\item{cell_type_label}{= the label for the cell types/cluster IDs to be used for calculating the module score}
}
\value{
A matrix of the module scores per cell types/cluster per cell.
}
\description{
This function calculates the module score for each cluster/cell type, using 
the Seurat AddModuleScore function.
}
