% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{Get_normalised_variables}
\alias{Get_normalised_variables}
\title{Get normalised variables}
\usage{
Get_normalised_variables(object, variate_names)
}
\arguments{
\item{object}{= seurat object}

\item{variate_names}{= the list of labels (from the meta.data of the seurat object)}
}
\value{
A matrix of the normalised variables per cell.
}
\description{
This function calculates the normalised variables for inclusion into the doublet detection 
model. This performs log-centred normalisation per sample
}
