% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{GetMitRibRatio}
\alias{GetMitRibRatio}
\title{Mitochondrial-ribosomal RNA ratio}
\usage{
GetMitRibRatio(object, mitrib = NULL, mito = NULL)
}
\arguments{
\item{object}{= seurat object}

\item{mitrib}{= ribosomal and mitichondrial gene IDs}

\item{mito}{= mitichondrial gene IDs}
}
\value{
An array of mitochondrial-ribosomal RNA ratios per cell.
}
\description{
This function calculates the mitochondrial-ribosomal RNA ratio. It assumes that the
contains the mitichondrial gene IDs begin with the prefix "MT-" and that the 
ribosomal gene IDs  begin with the prefix "RP-". If this is not the case, then
the gene IDs can be specified.
}
