% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{Predict_doublets}
\alias{Predict_doublets}
\title{Predict doublets}
\usage{
Predict_doublets(object, marker_array, doublets_training, variables_norm,
  training_trim = 1)
}
\arguments{
\item{object}{= seurat object}

\item{marker_array}{= a matrix of the module scores per cell types/cluster per cell.}

\item{doublets_training}{= an array of the training doublet/multiplet set (e.g. from VDJ-seq, CITE-seq or other).}

\item{variables_norm}{= a matrix of the normalised variables per cell.}

\item{training_trim}{= a parameter for filtering outliers from the training doublet/multiplet set (default = 1).}
}
\value{
doublet object (list of 1. the training doublet prediction table, 2. the doublet probabilities excluding the

training doublets, 3. the doublet probabilities including the training doublets, 4. the list of predicted doublets).
}
\description{
This function builds classifier based on "variate_names" plus "marker_array' (module scores 
for each cell type), and runs the classifier over the whole dataset. This assumes that the 
doublets_training are all cell IDs wihtin the seurat object.
}
